feat: change the path from assistcredit to role
